Factor conf parser out of address_map.lua
