+++ src/apps/lwaftr/address_map.lua
@@ -40,130 +40,11 @@ module(..., package.seeall)
 local ffi = require("ffi")
 local S = require("syscall")
 local rangemap = require("apps.lwaftr.rangemap")
+local Parser = require("apps.lwaftr.conf_parser").Parser
 
 local address_map_value = ffi.typeof([[
    struct { uint16_t psid_length; uint16_t shift; }
 ]])
-
-local Parser = {}
-
-function Parser.new(file)
-   local name = '<unknown>'
-   local mtime_sec, mtime_nsec
-   if type(file) == 'string' then
-      name, file = file, io.open(file)
-      -- Seems to be no way to fstat() the file.  Oh well.
-      local stat = S.stat(name)
-      mtime_sec, mtime_nsec = stat.st_mtime, stat.st_mtime_nsec
-      print(mtime_sec, mtime_nsec)
-   end
-   local ret = {
-      column=0, line=0, name=name, mtime_sec=mtime_sec, mtime_nsec=mtime_nsec
-   }
-   function ret.read_char() return file:read(1) end
-   function ret.cleanup() return file:close() end
-   ret.peek_char = ret.read_char()
-   return setmetatable(ret, {__index=Parser})
-end
-
-function Parser:error(msg, ...)
-   self.cleanup()
-   error(('%s:%d:%d: error: '..msg):format(self.name, self.line, self.column,
-                                           ...))
-end
-
-function Parser:next()
-   local chr = self.peek_char
-   if chr == '\n' then
-      self.column = 0
-      self.line = self.line + 1
-   elseif chr then
-      self.column = self.column + 1
-   end
-   self.peek_char = self.read_char()
-   return chr
-end
-
-function Parser:peek() return self.peek_char end
-function Parser:is_eof() return not self:peek() end
-
-function Parser:check(expected)
-   if self:is_eof() then
-      self:error("while looking for '%s', got EOF", expected)
-   elseif self:peek() == expected then
-      self:next()
-      return true
-   end
-   return false
-end
-
-function Parser:consume(expected)
-   if not self:check(expected) then
-      self:error("expected '%s', got '%s'", expected, self:peek())
-   end
-end
-
-function Parser:take_while(pattern)
-   local res = {}
-   while not self:is_eof() and self:peek():match(pattern) do
-      table.insert(res, self:next())
-   end
-   return table.concat(res)
-end
-
-function Parser:skip_whitespace() self:take_while('%s') end
-
-function Parser:parse_uint(min, max)
-   local tok = self:take_while('%d')
-   if tok == '' then self:error('expected a number') end
-   if #tok > #(tostring(max)) then
-      self:error('numeric constant too long: %s', tok)
-   end
-   local uint = tonumber(tok)
-   if uint < min or uint > max then
-      self:error('numeric constant out of range: %d', uint)
-   end
-   return uint
-end
-
-function Parser:parse_psid_param() return self:parse_uint(0, 16) end
-function Parser:parse_ipv4_quad() return self:parse_uint(0, 255) end
-
-function Parser:parse_kvlist(spec)
-   local res = {}
-   self:skip_whitespace()
-   self:consume('{')
-   self:skip_whitespace()
-   while not self:check('}') do
-      local key = self:take_while('[%w_]')
-      if key == '' then
-         self:error("expected a key=value property or a closing '}'")
-      end
-      if res[key] then self:error('duplicate key: %s', key) end
-      if not spec.parse[key] then self:error('unexpected key: %s', key) end
-      self:skip_whitespace()
-      self:consume('=')
-      self:skip_whitespace()
-      local val = spec.parse[key](self)
-      res[key] = val
-
-      -- Key-value pairs are separated by newlines or commas, and
-      -- terminated by }.  A trailing comma is optional.
-      local line = self.line
-      self:skip_whitespace()
-      local has_comma = self:check(',')
-      if has_comma then self:skip_whitespace() end
-      if self:check('}') then break end
-      if not has_comma and self.line == line then
-         self:error('properties should be separated by commas or newlines')
-      end
-   end
-   for k, default in pairs(spec.defaults) do
-      if not res[k] then res[k] = default(res) end
-   end
-   spec.validate(self, res)
-   return res
-end
 
 local psid_info_spec = {
    parse={
@@ -174,76 +55,30 @@ local psid_info_spec = {
       psid_length=function(config) return 16 - (config.shift or 16) end,
       shift=function(config) return 16 - (config.psid_length or 0) end
    },
-   validate=function(self, config)
+   validate=function(parser, config)
       if config.psid_length + config.shift ~= 16 then
-         self:error('psid_length %d + shift %d should add up to 16',
-                    config.psid_length, config.shift)
+         parser:error('psid_length %d + shift %d should add up to 16',
+                      config.psid_length, config.shift)
       end
    end
 }
 
-function Parser:parse_psid_info()
-   return self:parse_kvlist(psid_info_spec)
-end
-
--- Parse IPv4 address as host-endian integer.
-function Parser:parse_ipv4()
-   local q1 = self:parse_ipv4_quad()
-   self:consume('.')
-   local q2 = self:parse_ipv4_quad()
-   self:consume('.')
-   local q3 = self:parse_ipv4_quad()
-   self:consume('.')
-   local q4 = self:parse_ipv4_quad()
-   return q1*2^24 + q2*2^16 + q3*2^8 + q4
-end
-
-function Parser:parse_ipv4_range()
-   local range_begin, range_end
-   range_begin = self:parse_ipv4()
-   self:skip_whitespace()
-   if self:check('-') then
-      self:skip_whitespace()
-      range_end = self:parse_ipv4()
-   else
-      range_end = range_begin
-   end
-   if range_end < range_begin then
-      self:error('invalid IPv4 address range (end before begin)')
-   end
-   return { min=range_begin, max=range_end }
-end
-
-function Parser:parse_ipv4_range_list()
-   local ranges = {}
-   repeat
-      self:skip_whitespace()
-      table.insert(ranges, self:parse_ipv4_range())
-      self:skip_whitespace()
-   until not self:check(',')
-   return ranges
-end
-
-function Parser:parse_entry()
-   local range_list = self:parse_ipv4_range_list()
-   local info = self:parse_psid_info()
-   info.range_list = range_list
-   return info
-end
-
-function Parser:parse_entries()
+function parse_address_map(parser)
    local entries = {}
-   self:skip_whitespace()
-   while not self:is_eof() do
-      table.insert(entries, self:parse_entry())
-      self:skip_whitespace()
+   parser:skip_whitespace()
+   while not parser:is_eof() do
+      local range_list = parser:parse_ipv4_range_list()
+      local info = parser:parse_property_list(psid_info_spec, '{', '}')
+      info.range_list = range_list
+      table.insert(entries, info)
+      parser:skip_whitespace()
    end
-   self.cleanup()
+   parser.cleanup()
    return entries
 end
 
 local function parse(stream)
-   return Parser.new(stream):parse_entries()
+   return parse_address_map(Parser.new(stream))
 end
 
 local function attach_lookup_helper(map)
